refactor(scripts): share mobile changelog extraction
