+++ scripts/lib/android-version.ts
@@ -3,6 +3,7 @@ import { readFileSync } from "node:fs";
 import path from "node:path";
 import { decodeIosAppStoreVersion } from "./ios-release-plan.ts";
 import { encodeIosAppStoreVersion } from "./ios-version.ts";
+import { extractChangelogSection } from "./mobile-changelog.ts";
 import { readMobileVersionManifest } from "./mobile-version.ts";
 import { parsePinnedReleaseVersion } from "./release-version.mjs";
 
@@ -153,33 +154,6 @@ function renderLegacyAndroidVersionProperties(
   version: Pick<ResolvedAndroidVersion, "canonicalVersion" | "versionCode">,
 ): string {
   return `# Shared Android version defaults.\n# Source of truth: apps/android/version.json\n# Generated by scripts/android-sync-versioning.ts.\n\nOPENCLAW_ANDROID_VERSION_NAME=${version.canonicalVersion}\nOPENCLAW_ANDROID_VERSION_CODE=${version.versionCode}\n`;
-}
-
-function matchChangelogHeading(line: string, heading: string): boolean {
-  const normalized = line.trim();
-  return normalized === `## ${heading}` || normalized.startsWith(`## ${heading} - `);
-}
-
-export function extractChangelogSection(content: string, heading: string): string | null {
-  const lines = content.split(/\r?\n/u);
-  const startIndex = lines.findIndex((line) => matchChangelogHeading(line, heading));
-  if (startIndex === -1) {
-    return null;
-  }
-
-  let endIndex = lines.length;
-  for (let index = startIndex + 1; index < lines.length; index += 1) {
-    if (lines[index]?.startsWith("## ")) {
-      endIndex = index;
-      break;
-    }
-  }
-
-  const body = lines
-    .slice(startIndex + 1, endIndex)
-    .join("\n")
-    .trim();
-  return body || null;
 }
 
 export function renderAndroidReleaseNotes(
