+++ scripts/lib/ios-version.ts
@@ -1,6 +1,7 @@
 // Ios Version script supports OpenClaw repository automation.
 import { readFileSync } from "node:fs";
 import path from "node:path";
+import { extractChangelogSection } from "./mobile-changelog.ts";
 import { mobileVersionPath, readMobileVersionManifest } from "./mobile-version.ts";
 import { parsePinnedReleaseVersion, parseReleaseVersion } from "./release-version.mjs";
 
@@ -111,33 +112,6 @@ export function resolveIosVersion(
     versionSource: explicitReleaseVersion ? "explicit" : "mobile",
     versionSourcePath: explicitReleaseVersion ? null : mobileVersionPath(rootDir),
   };
-}
-
-function matchChangelogHeading(line: string, heading: string): boolean {
-  const normalized = line.trim();
-  return normalized === `## ${heading}` || normalized.startsWith(`## ${heading} - `);
-}
-
-export function extractChangelogSection(content: string, heading: string): string | null {
-  const lines = content.split(/\r?\n/);
-  const startIndex = lines.findIndex((line) => matchChangelogHeading(line, heading));
-  if (startIndex === -1) {
-    return null;
-  }
-
-  let endIndex = lines.length;
-  for (let index = startIndex + 1; index < lines.length; index += 1) {
-    if (lines[index]?.startsWith("## ")) {
-      endIndex = index;
-      break;
-    }
-  }
-
-  const body = lines
-    .slice(startIndex + 1, endIndex)
-    .join("\n")
-    .trim();
-  return body || null;
 }
 
 export function renderIosReleaseNotes(
